medseg/server/requirement.txt
Flask==1.1.1
tensorboard==2.0.2
tensorflow==2.0.0
tensorflow-estimator==2.0.1
Keras==2.3.1
matplotlib==3.1.1
python-dotenv==0.10.3
nibabel==2.5.1
nilearn==0.5.2
SimpleITK==1.2.3
git+https://www.github.com/farizrahman4u/keras-contrib.git